+++ test/usecases/hash_serialization_test.rb
@@ -54,4 +54,48 @@ class HashSerializationTest < Minitest::Test
       InstanceMethodHashResource.new(hash).serialize
     )
   end
+
+  class HashWithMissingKeysResource
+    include Alba::Resource
+
+    attributes :id, :name, :email
+  end
+
+  def test_hash_serialization_with_missing_keys
+    hash = {id: 1, name: 'test'}
+    assert_equal(
+      '{"id":1,"name":"test","email":null}',
+      HashWithMissingKeysResource.new(hash).serialize
+    )
+  end
+
+  class HashLike < Hash
+    def fetch(key, *args)
+      super(key.to_s, *args)
+    end
+  end
+
+  def test_hash_like_serialization_with_missing_keys
+    hash = HashLike['id' => 1, 'name' => 'test'] # rubocop:disable Style/StringHashKeys
+    assert_equal(
+      '{"id":1,"name":"test","email":null}',
+      HashWithMissingKeysResource.new(hash).serialize
+    )
+  end
+
+  class PreferObjectMethodHashResource
+    include Alba::Resource
+
+    prefer_object_method!
+
+    attributes :id, :name, :email
+  end
+
+  def test_hash_serialization_with_missing_keys_prefer_object_method
+    hash = {id: 1, name: 'test'}
+    assert_equal(
+      '{"id":1,"name":"test","email":null}',
+      PreferObjectMethodHashResource.new(hash).serialize
+    )
+  end
 end